backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
sympy
python-dotenv
pillow
motor  
certifi
easyocr
opencv-python
numpy